+++ example/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - cesium-wind-layer@0.7.0
+
 ## 0.4.0
 
 ### Minor Changes